[AccessBuilder] Move from String to Name
After the version bump to v8.11.1 some tests ran into segfaults
when accessing taint on strings. This is caused by new code that
allocates ConsStrings and yet wan't patched to initialize the
taint pointer as a nullptr. This adds correct initialization of
the taint pointer for the new code that allocates ConsStrings.